+++ src/components/biz/FormModal/formModal.type.ts
@@ -6,7 +6,7 @@ import {
   IFormActionsBase,
 } from "../../../components";
 import { ComputedRef, VNodeTypes } from "vue";
-import { Fn, IObj } from "../../../types";
+import { IObj } from "../../../types";
 
 export interface IFormModalProp extends IModalProps, IFormProps {
   // 初始值
@@ -31,7 +31,7 @@ export interface IFormModalProp extends IModalProps, IFormProps {
   pageChange: (obj?: { current?: number; pageSize?: number }) => Promise<void>;
 }
 
-export interface IFormModalActions extends IModalActionBase, IFormActionsBase, ICommonAction<IFormModalProp> {
+export interface IFormModalActions extends ICommonAction<IFormModalProp> {
   // 手动打开新增弹框
   openNew: (i?: IObj) => Promise<void>;
   // 手动打开修改弹框
@@ -44,11 +44,7 @@ export interface IFormModalActions extends IModalActionBase, IFormActionsBase, I
   openAddModal: (record?: any) => void;
   // 手动触发传入的openEditModal
   openEditModal: (param: any) => void;
-  // 继承的
-  openModal: () => Promise<any>;
-  closeModal: () => Promise<any>;
-  openLoading: () => void;
-  closeLoading: () => void;
-  setOk: Fn;
-  setCancel: Fn;
+
+  modalMethods: IModalActionBase;
+  formMethods: IFormActionsBase;
 }