+++ src/components/common/table/hooks/index.ts
@@ -3,4 +3,4 @@ export { default as useDataSource } from "./useDataSource";
 export { default as useRowSelection } from "./useRowSelection";
 export { default as useTableScroll } from "./useTableScroll";
 export { default as useTableStyle } from "./useTableStyle";
-export { tableMethods } from "./useTable";
+export { getTableMethods } from "./useTable";