+++ changelog.md
@@ -14,7 +14,13 @@
 - basicModal 点击事件 2 秒内只能触发一次
 
 ### 1.1.11
-- 对formModal新增了basicModal的基本操作，来源于继承
+
+- 对 formModal 新增了 basicModal 的基本操作，来源于继承
 
 ### 1.1.12
-- 对formModal新增了basicForm的基本操作，来源于继承
+
+- 对 formModal 新增了 basicForm 的基本操作，来源于继承
+
+### 1.1.13
+
+- 对 bizTable 和 formModal 的继承进行修改，actions 上不在暴露那么多，只暴露对象